#83 Implementacao de backend
* Adicao de testes para repo e sources
* adicao de sources para recipe, step e ingredient

+++ projetos_finais/Baking_app/Projeto_App/app/src/main/java/com/alex/baking/app/data/model/Ingredient.java
@@ -2,15 +2,15 @@ package com.alex.baking.app.data.model;
 
 public class Ingredient extends BaseModel {
 
-	private Long quantity;
+	private Double quantity;
 	private MeasureType measure;
 	private String ingredient;
 
-	public Long getQuantity() {
+	public Double getQuantity() {
 		return quantity;
 	}
 
-	public void setQuantity(Long quantity) {
+	public void setQuantity(Double quantity) {
 		this.quantity = quantity;
 	}
